refactor(web): migrate repository detail page to useSuspenseQueries
Replace useQuery with useSuspenseQueries to improve data fetching performance and eliminate waterfall loading. Extract data fetching logic into a separate RepositoryContent component wrapped with Suspense boundary.

Changes:
- Create -repository-content.tsx component with useSuspenseQueries
- Fetch repository, branches, and pulls data in parallel
- Add Suspense boundary with loading fallback
- Remove manual loading and error state handling
- Simplify component structure
